+++ health/micro-ui/web/micro-ui-internals/packages/modules/campaign-manager/src/components/XlsPreview.js
@@ -1,6 +1,8 @@
 import { PopUp, SVG, DownloadIcon } from "@egovernments/digit-ui-react-components";
 import React from "react";
-import DocViewer, { DocViewerRenderers } from "@cyntler/react-doc-viewer";
+// Removed @cyntler/react-doc-viewer to fix CVE-2024-4367 (pdfjs-dist vulnerability).
+// Replaced with direct Office Online iframe which provides the same XLSX preview functionality.
+// import DocViewer, { DocViewerRenderers } from "@cyntler/react-doc-viewer";
 import { Button } from "@egovernments/digit-ui-components";
 import { useTranslation } from "react-i18next";
 import { PRIMARY_COLOR } from "../utils";
@@ -17,15 +19,15 @@ const ArrowBack = ({ className = "", height = "15", width = "15", styles = {} })
 };
 function XlsPreview({ file, ...props }) {
   const { t } = useTranslation();
-  const documents = file
-    ? [
-      {
-        fileType: "xlsx",
-        fileName: file?.filename,
-        uri: file?.url,
-      },
-    ]
-    : null;
+  // const documents = file
+  //   ? [
+  //     {
+  //       fileType: "xlsx",
+  //       fileName: file?.filename,
+  //       uri: file?.url,
+  //     },
+  //   ]
+  //   : null;
 
   return (
     <PopUp className="campaign-data-preview" style={{ flexDirection: "column" }}>
@@ -46,7 +48,8 @@ function XlsPreview({ file, ...props }) {
         />
       </div>
       <div className="campaign-popup-module" style={{ marginTop: "1.5rem" }}>
-        <DocViewer
+        {/* Replaced DocViewer with direct Office Online iframe to fix CVE-2024-4367 */}
+        {/* <DocViewer
           style={{ height: "80vh", overflowY: "hidden" }}
           theme={{
             primary: PRIMARY_COLOR,
@@ -59,6 +62,11 @@ function XlsPreview({ file, ...props }) {
           }}
           documents={documents}
           pluginRenderers={DocViewerRenderers}
+        /> */}
+        <iframe
+          src={`https://view.officeapps.live.com/op/embed.aspx?src=${encodeURIComponent(file?.url)}`}
+          style={{ width: "100%", height: "80vh", border: "none" }}
+          title="XLSX Preview"
         />
       </div>
     </PopUp>